+++ src/argilla/server/schemas/v1/fields.py
@@ -13,18 +13,51 @@
 #  limitations under the License.
 
 from datetime import datetime
-from typing import Literal, Optional
+from typing import Annotated, List, Literal, Optional
 from uuid import UUID
 
+from typing_extensions import Annotated
+
 from argilla.server.enums import FieldType
-from argilla.server.pydantic_v1 import BaseModel
+from argilla.server.pydantic_v1 import BaseModel, constr
+from argilla.server.pydantic_v1 import Field as PydanticField
 from argilla.server.schemas.base import UpdateSchema
-from argilla.server.schemas.v1.datasets import FieldTitle
+
+FIELD_CREATE_NAME_REGEX = r"^(?=.*[a-z0-9])[a-z0-9_-]+$"
+FIELD_CREATE_NAME_MIN_LENGTH = 1
+FIELD_CREATE_NAME_MAX_LENGTH = 200
+
+FIELD_CREATE_TITLE_MIN_LENGTH = 1
+FIELD_CREATE_TITLE_MAX_LENGTH = 500
+
+
+FieldName = Annotated[
+    constr(
+        regex=FIELD_CREATE_NAME_REGEX,
+        min_length=FIELD_CREATE_NAME_MIN_LENGTH,
+        max_length=FIELD_CREATE_NAME_MAX_LENGTH,
+    ),
+    PydanticField(..., description="The name of the field"),
+]
+
+
+FieldTitle = Annotated[
+    constr(
+        min_length=FIELD_CREATE_TITLE_MIN_LENGTH,
+        max_length=FIELD_CREATE_TITLE_MAX_LENGTH,
+    ),
+    PydanticField(..., description="The title of the field"),
+]
 
 
 class TextFieldSettings(BaseModel):
     type: Literal[FieldType.text]
     use_markdown: bool = False
+
+
+class TextFieldSettingsUpdate(UpdateSchema):
+    type: Literal[FieldType.text]
+    use_markdown: bool
 
 
 class Field(BaseModel):
@@ -41,9 +74,15 @@ class Field(BaseModel):
         orm_mode = True
 
 
-class TextFieldSettingsUpdate(UpdateSchema):
-    type: Literal[FieldType.text]
-    use_markdown: bool
+class Fields(BaseModel):
+    items: List[Field]
+
+
+class FieldCreate(BaseModel):
+    name: FieldName
+    title: FieldTitle
+    required: Optional[bool]
+    settings: TextFieldSettings
 
 
 class FieldUpdate(UpdateSchema):
