+++ tests/unit/server/api/v1/test_fields.py
@@ -19,7 +19,7 @@ from uuid import uuid4
 import pytest
 from argilla.server.constants import API_KEY_HEADER_NAME
 from argilla.server.models import DatasetStatus, Field, UserRole
-from argilla.server.schemas.v1.datasets import FIELD_CREATE_TITLE_MAX_LENGTH
+from argilla.server.schemas.v1.fields import FIELD_CREATE_TITLE_MAX_LENGTH
 from sqlalchemy import func, select
 
 from tests.factories import (
